+++ src/exactly_lib/cli/program_modes/symbol/impl/reports/symbol_info.py
@@ -10,20 +10,40 @@ from exactly_lib.type_system.value_type import ValueType
 SYMBOL_INFO = TypeVar('SYMBOL_INFO')
 
 
+class SourceInfo:
+    def __init__(self,
+                 source_location_info: Optional[SourceLocationInfo],
+                 source_lines: Optional[List[str]]):
+        self.source_location_info = source_location_info
+        self.source_lines = source_lines
+
+    @staticmethod
+    def of_location_info(location_info: SourceLocationInfo):
+        return SourceInfo(location_info, None)
+
+    @staticmethod
+    def of_lines(lines: List[str]):
+        return SourceInfo(None, lines)
+
+    @staticmethod
+    def empty():
+        return SourceInfo(None, None)
+
+
 class ContextAnd(Generic[SYMBOL_INFO]):
     def __init__(self,
                  phase: Phase,
-                 source_location_info: Optional[SourceLocationInfo],
+                 source_info: SourceInfo,
                  value: SYMBOL_INFO):
         self._phase = phase
-        self._source_location_info = source_location_info
+        self._source_info = source_info
         self._value = value
 
     def phase(self) -> Phase:
         return self._phase
 
-    def source_location_info(self) -> Optional[SourceLocationInfo]:
-        return self._source_location_info
+    def source_info(self) -> SourceInfo:
+        return self._source_info
 
     def value(self) -> SYMBOL_INFO:
         return self._value