+++ src/exactly_lib/cli/program_modes/symbol/impl/symbol_usage_resolving.py
@@ -1,13 +1,14 @@
 import itertools
 
-from typing import List, Sequence, Iterator, Dict, Callable, Optional, Any
+from typing import List, Sequence, Iterator, Dict, Callable, Optional
 
 from exactly_lib.cli.program_modes.symbol.impl.reports.symbol_info import SYMBOL_INFO, SymbolDefinitionInfo, \
-    DefinitionsResolver, ContextAnd
+    DefinitionsResolver, ContextAnd, SourceInfo
 from exactly_lib.symbol.symbol_usage import SymbolDefinition, SymbolUsage, SymbolReference, SymbolUsageVisitor
 from exactly_lib.test_case import phase_identifier
 from exactly_lib.test_case.phase_identifier import Phase
 from exactly_lib.test_case.phases import setup, before_assert, assert_, cleanup
+from exactly_lib.test_case.phases.act import ActPhaseInstruction
 from exactly_lib.test_case.test_case_doc import TestCaseOfInstructions2, ElementWithSourceLocation
 
 
@@ -25,9 +26,7 @@ class DefinitionsInfoResolverFromTestCase(DefinitionsResolver):
         def mk_definition(definition: ContextAnd[SymbolDefinition]) -> SymbolDefinitionInfo:
             name = definition.value().name
 
-            references_to_symbol = []
-            if name in references:
-                references_to_symbol = references[name]
+            references_to_symbol = references.get(name, [])
 
             return SymbolDefinitionInfo(definition.phase(),
                                         definition.value(),
@@ -39,15 +38,10 @@ class DefinitionsInfoResolverFromTestCase(DefinitionsResolver):
                    )
 
     def symbol_usages(self) -> Iterator[ContextAnd[SymbolUsage]]:
-        def mk_act_phase_sym_usage(usage: SymbolUsage) -> ContextAnd[SymbolUsage]:
-            return ContextAnd(phase_identifier.ACT,
-                              None,
-                              usage)
-
         return itertools.chain.from_iterable([
             _symbol_usages_from(phase_identifier.SETUP,
                                 self.test_case.setup_phase, setup.get_symbol_usages),
-            map(mk_act_phase_sym_usage, self.act_phase),
+            self._act_phase_symbol_usages(),
             _symbol_usages_from(phase_identifier.BEFORE_ASSERT,
                                 self.test_case.before_assert_phase,
                                 before_assert.get_symbol_usages),
@@ -57,19 +51,50 @@ class DefinitionsInfoResolverFromTestCase(DefinitionsResolver):
                                 self.test_case.cleanup_phase, cleanup.get_symbol_usages),
         ])
 
+    def _act_phase_symbol_usages(self) -> Iterator[ContextAnd[SymbolUsage]]:
+        usages_extractor = _UsagesExtractor()
+        source_info = self._act_phase_source_info()
+
+        def mk_act_phase_sym_usage(usage: SymbolUsage) -> ContextAnd[SymbolUsage]:
+            return ContextAnd(phase_identifier.ACT,
+                              source_info,
+                              usage)
+
+        return map(
+            mk_act_phase_sym_usage,
+            itertools.chain.from_iterable([
+                usages_extractor.visit(symbol_usage)
+                for symbol_usage in self.act_phase
+            ])
+        )
+
+    def _act_phase_source_info(self) -> SourceInfo:
+        def instruction_source_lines(instruction: ElementWithSourceLocation[ActPhaseInstruction]) -> Sequence[str]:
+            return instruction.value.source_code().lines
+
+        source_lines = list(
+            itertools.chain.from_iterable(
+                map(instruction_source_lines, self.test_case.act_phase)
+            )
+        )
+
+        if source_lines:
+            return SourceInfo.of_lines(source_lines)
+        else:
+            return SourceInfo.empty()
+
     @staticmethod
     def references(usages: List[ContextAnd[SymbolUsage]]
                    ) -> Dict[str, List[ContextAnd[SymbolReference]]]:
-        references = itertools.chain.from_iterable(
-            map(_extract_symbol_references, usages)
+        context_and_reference_iter = filter(
+            is_not_none,
+            map(_extract_symbol_reference, usages)
         )
         ret_val = {}
-        for reference in references:
-            name = reference.value().name
-            if name in ret_val:
-                ret_val[name].append(reference)
-            else:
-                ret_val[name] = [reference]
+        for context_and_reference in context_and_reference_iter:
+            name = context_and_reference.value().name
+            refs_for_name = ret_val.setdefault(name, [])
+            refs_for_name.append(context_and_reference)
 
         return ret_val
 
@@ -78,13 +103,21 @@ def _symbol_usages_from(phase: Phase,
                         elements: Sequence[ElementWithSourceLocation[SYMBOL_INFO]],
                         symbol_usages_getter: Callable[[SYMBOL_INFO], Sequence[SymbolUsage]]
                         ) -> Iterator[ContextAnd[SymbolUsage]]:
+    usages_extractor = _UsagesExtractor()
+
+    def get_direct_and_indirect_symbol_usages(symbol_info: SYMBOL_INFO) -> Iterator[SymbolUsage]:
+        return itertools.chain.from_iterable(
+            usages_extractor.visit(symbol_usage)
+            for symbol_usage in symbol_usages_getter(symbol_info)
+        )
+
     symbol_usages_sequence_list = [
         [
             ElementWithSourceLocation(
                 element.source_location_info,
                 symbol_usage
             )
-            for symbol_usage in symbol_usages_getter(element.value)
+            for symbol_usage in get_direct_and_indirect_symbol_usages(element.value)
 
         ]
         for element in elements
@@ -92,7 +125,7 @@ def _symbol_usages_from(phase: Phase,
 
     def mk_item(element: ElementWithSourceLocation[SymbolUsage]) -> ContextAnd[SymbolUsage]:
         return ContextAnd(phase,
-                          element.source_location_info,
+                          SourceInfo.of_location_info(element.source_location_info),
                           element.value)
 
     return map(mk_item,
@@ -104,36 +137,31 @@ def _extract_symbol_definition(usage: ContextAnd[SymbolUsage]) -> Optional[Conte
     value = usage.value()
     if isinstance(value, SymbolDefinition):
         return ContextAnd(usage.phase(),
-                          usage.source_location_info(),
+                          usage.source_info(),
                           value)
     else:
         return None
 
 
-def _extract_symbol_references(usage: ContextAnd[SymbolUsage]) -> List[ContextAnd[SymbolReference]]:
-    return _ReferencesExtractor(usage).visit(usage.value())
+def _extract_symbol_reference(context_and_usage: ContextAnd[SymbolUsage]) -> Optional[ContextAnd[SymbolReference]]:
+    usage = context_and_usage.value()
+    if isinstance(usage, SymbolReference):
+        return ContextAnd(context_and_usage.phase(),
+                          context_and_usage.source_info(),
+                          usage)
+    else:
+        return None
 
 
 def is_not_none(x) -> bool:
     return x is not None
 
 
-class _ReferencesExtractor(SymbolUsageVisitor):
-    def __init__(self, context: ContextAnd[Any]):
-        self._context = context
+class _UsagesExtractor(SymbolUsageVisitor):
+    def _visit_definition(self, definition: SymbolDefinition) -> List[SymbolUsage]:
+        return [definition] + list(definition.references)
 
-    def _visit_definition(self, definition: SymbolDefinition) -> List[ContextAnd[SymbolReference]]:
+    def _visit_reference(self, reference: SymbolReference) -> List[SymbolUsage]:
         return [
-            self._of(reference)
-            for reference in definition.references
+            reference
         ]
-
-    def _visit_reference(self, reference: SymbolReference) -> List[ContextAnd[SymbolReference]]:
-        return [
-            self._of(reference)
-        ]
-
-    def _of(self, reference: SymbolReference) -> ContextAnd[SymbolReference]:
-        return ContextAnd(self._context.phase(),
-                          self._context.source_location_info(),
-                          reference)