+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [next_release] - 201?-??-??
+
+### Fixed
+
+ - symbol reference reporting: Source location
+ 
 ## [0.9.1] - 2019-01-17
 
 ### Added